refactor: extract TrackerExportFingerprint.swift from ExportReceiptStore.swift
Extracts the `TrackerExportFingerprint` enum (FNV-1a hash helper for
tracker export IDs, ~30 lines) from ExportReceiptStore.swift into its
own file. ExportReceiptStore.swift retains only the `actor
ExportReceiptStore` implementation.

Byte-identical move of lines 159-188. Only imports Foundation.

ExportReceiptStore.swift: 188 → 157 lines. Tests: 667.
